Move chart to its own component
